requirements.txt
langchain>=0.1.0
langchain-core>=0.1.10
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-community>=0.0.13
langchain-google-genai>=0.0.5
tavily-python>=0.3.1
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0 